requirements.txt
pandas==2.2.3
requests==2.32.3
SQLAlchemy==2.0.37
psycopg==3.2.4
psycopg-binary==3.2.4
python-dotenv==1.0.1